LayoutTests/fast/forms/date-multiple-fields/date-multiple-fields-focus-expected.txt

Check if element.focus() does not focus on disabled sub-fields.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.youngestShadowRoot(dateInput).activeElement.getAttribute("pseudo") is "-webkit-datetime-edit-month-field"
PASS internals.youngestShadowRoot(dateInput).activeElement.getAttribute("pseudo") is "-webkit-datetime-edit-month-field"
PASS successfullyParsed is true

TEST COMPLETE
